% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{albedo_adjust}
\alias{albedo_adjust}
\title{Corrects albedo derived from aerial photographs}
\usage{
albedo_adjust(photoalbedo, modisalbedo)
}
\arguments{
\item{photoalbedo}{a SpatRaster of albedo derived using [albedo_fromaerial()].}

\item{modisalbedo}{a SpatRaster of albedo derived from MODIS data using [albedo_download()]
and [albedo_process()] usually of much coarser resoltuion than `photoalbedo`.}
}
\value{
a SpatRaster of adjusted albedo
}
\description{
Using MODIS, derived albedo, corrects albedo derived using function
[albedo_fromaerial()] to correct for brightness and contrast in the aerial image
}
