% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{albedo_download}
\alias{albedo_download}
\title{Download MODIS albedo data}
\usage{
albedo_download(r, tme, pathout, credentials)
}
\arguments{
\item{r}{a SpatRaster with a coordinate reference system defined. Used
to the location for which to download data.}

\item{tme}{POSIXlt object covering time period for which data are required
(see details).}

\item{pathout}{directory to which data will be downloaded.}

\item{credentials}{a data.frame of user credentials (see vignette for
setting this up)}

\item{pathtopython}{directory in which Python is saved (see
vignette instructions for installing and setting up Python)}
}
\description{
Downloads 500m grid resolution MODIS albedo data from LPDAAC_ECS.
}
\details{
500 m grid resolution albedo data derived from MODIS imagery are downloaded
from LPDAAC_ECS using the `luna` package. Data are available globally from 16th
February 2000 onward. Multiple files are downloaded and stored in the directory
given by `pathout`. Use [albedo_process()] to tile these together into one file.
See vignette for examples.
}
\seealso{
[albedo_process()], [albedo_fromaerial()], [albedo_adjust()]
}
