% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{albedo_fromaerial}
\alias{albedo_fromaerial}
\title{Derives albedo from aerial photographs}
\usage{
albedo_fromaerial(
  RGBimage,
  CIRimage,
  RGBbandmins = c(620, 495, 450),
  RGBbandmaxs = c(750, 570, 495),
  CIRbandmins = c(750, 620, 495),
  CIRbandmaxs = c(900, 750, 570)
)
}
\arguments{
\item{RGBimage}{a 3-layer red-green-blue SpatRaster with a coordinate reference system
defined and at least partially over-lapping with CIRimage.}

\item{CIRimage}{a 3-layer colour-infra-red SpatRaster with a coordinate reference system
defined and at least partially over-lapping with RGBimage.}

\item{RGBbandmins}{minimum wavelengths of bands in RGBimage (nm).
Default assumes image ordered Red-Green-Blue}

\item{RGBbandmaxs}{maximum wavelengths of bands in RGBimage (nm).
Default assumes image ordered Red-Green-Blue}

\item{CIRbandmins}{minimum wavelengths of bands in CIRimage (nm).
Default assumes image ordered Near-infrared, Red, Green}

\item{CIRbandmaxs}{maximum wavelengths of bands in CIRimage (nm).
Default assumes image ordered Near-infrared, Red, Green}
}
\value{
a SpatRaster of albedo
}
\description{
Derives albedo from red-green-blue and colour-infrared aerial photographs
}
\details{
derives albedo by weighting each image band by the band width and spectral
density of light emitted from the sun in that band range. Note that technically this
function does not derive true albedo as the brightness and contrast of the aerial
image may have been artificially altered during processing. use [albedo_adjust()] to
deal with this  problem
}
\seealso{
[albedo_adjust()]
}
