% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{create_soilgrid}
\alias{create_soilgrid}
\title{Create vegetation inputs for grid model}
\usage{
create_soilgrid(soildata, refldata, landcover, water = 80)
}
\arguments{
\item{soildata}{a multilayer SpatRaster of soil data as returned by [soildata_download()].}

\item{refldata}{a list of ground and leaf reflectances as returned by [reflectance_calc()].}

\item{landcover}{a SpatRaster of landcover class as returned by [lcover_download()].}

\item{water}{integer value in `landcover` corresponding to water (e.g. 80 for ESA dataset)}
}
\value{
an object of class `soilcharac`, namely a list of the following objects:
\describe{
  \item{soiltype}{a PackedSpatRaster of soil types specified using integer values}
  \item{groundr}{a PackedSpatRaster object of ground reflectances}
  \item{rho}{a PackedSpatRaster object of bulk densities (Mg / m^3)}
  \item{Vm}{a PackedSpatRaster object of the volumetric mineral fraction of soil}
  \item{Vq}{a PackedSpatRaster object of the volumetric quartz fraction of soil}
  \item{Mc}{a PackedSpatRaster object of the mass fraction of clay}
}
}
\description{
Creates an object of class vegparams used as an input to `microclimf`
or `microgrid`.
}
