% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credentials}
\alias{credentials}
\title{A template data frame for storing user credentials}
\format{
a data frame with the following elements:
\describe{
 \item{Site}{Repository for which user credentials are being provided}
 \item{username}{username}
 \item{password}{password}
 \item{Datasets}{Descriptor of datasets accessed from that repository}
 \item{Weblink}{Weblink for registration}
}
}
\usage{
credentials
}
\description{
A data frame to be used as a template for storing user credentials
}
\keyword{datasets}
