% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{dem_download}
\alias{dem_download}
\title{Download Digital Elevation Data}
\usage{
dem_download(r, msk = TRUE, zeroasna = FALSE)
}
\arguments{
\item{r}{a SpatRaster object used as a template for downloading data with an EPSG code defined.}

\item{msk}{optional logical indicating whether to apply `r` as a mask to downloaded
data (see details),}

\item{zeroasna}{optional logical indicating whether to set values wiht an elevation of zero
(e.g. sea) as NA. Generally not needed if `msk = TRUE` (see details).}
}
\value{
a SpatRaster of digital elevation data
}
\description{
Downloads Digital Elevation Data from the Amazon Web Server
}
\details{
the data returned will match the resolution, extent and coordinate reference
system of `r`. The resolution of available data varies globally however, so in some instances
it may be, that after resampling to match `r` the resulting data may be derived by
interpolation. A map showing the resolution of available data is available from
https://github.com/tilezen/joerd/blob/master/docs/images/footprints-preview.png.
For use with or use with microclimate packages, the returned SpatRaster, and hence
`r` should have a Mercator type Coordinate Reference System such that X, y and Z
units are all in metres. The data available from AWS blend digital elevation and
bathymetry data, the latter generally only available at coarser resolution. This
can cause oddities in coastal regions, where it is sensible to use `r` as a land-sea
mask by setting `msk` to `TRUE`.Alternatively by setting `zeroasna` to `TRUE` anything
below sea-level is set to `NA`.
}
