% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.altcorrectp}
\alias{.altcorrectp}
\title{applies altitudinal correction to climate data for a point}
\usage{
.altcorrectp(climdata, dtmc_p, elev_p, altcor = 2)
}
\arguments{
\item{climdata}{a dataframe of weather variables as output by `climpoint_extract`}

\item{dtmc}{a raster of climate data}
}
\description{
This function applies a lapse rate correction to the climdata
dataframe as used by the `micropoint` package based on the difference between the
elevation of the point and the elevation of the era5 data
}
