% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.biascorrect}
\alias{.biascorrect}
\title{apply bias correction to UKCP18 data}
\usage{
.biascorrect(hist_obs, hist_mod, fut_mod, rangelims = 1.05, silent = FALSE)
}
\description{
apply bias correction to UKCP18 data
}
