% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.calcmu}
\alias{.calcmu}
\title{Calculates diurnal temperature range adjustment for NCEP coastal grid cells}
\usage{
.calcmu(lat, lon)
}
\description{
Calculates diurnal temperature range adjustment for NCEP coastal grid cells
}
