% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.correct_ukcp_dates}
\alias{.correct_ukcp_dates}
\title{correct dates form downloaded ukcp18 nc file}
\usage{
.correct_ukcp_dates(ukcp_dates)
}
\description{
correct dates form downloaded ukcp18 nc file
}
