% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.extract_clima}
\alias{.extract_clima}
\title{equivelent of mcera5::extract_clima but corrected}
\usage{
.extract_clima(nc, r, resampleout = FALSE)
}
\arguments{
\item{nc}{vector of two filepaths. The first is the path to era5 reanalysis
download and the second is the path to era5 land download}

\item{r}{template spatRaster}

\item{resampleout}{should the output be resampled to the spatial resolution
of r}
}
\description{
equivelent of mcera5::extract_clima but corrected
}
