% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.extractvar}
\alias{.extractvar}
\title{Extracts a NCEP variable from a SpatRaster returtened by get_ncep}
\usage{
.extractvar(cr, ll)
}
\description{
Extracts a NCEP variable from a SpatRaster returtened by get_ncep
}
