% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.fill_calendar_data}
\alias{.fill_calendar_data}
\title{convert downloaded ukcp18 nc from 360 day to 365 day}
\usage{
.fill_calendar_data(
  ukcp_r,
  real_dates,
  testplot = FALSE,
  plotdays = c(89:91, 242:244)
)
}
\description{
convert downloaded ukcp18 nc from 360 day to 365 day
}
