% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.find_ukcp_decade}
\alias{.find_ukcp_decade}
\title{get the right UKCP decade}
\usage{
.find_ukcp_decade(collection = c("land-gcm", "land-rcm"), startdate, enddate)
}
\description{
get the right UKCP decade
}
