% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.gedi_distweight}
\alias{.gedi_distweight}
\title{Calculate the shape of the average vertical profile for a given PAI range and apply the shape to the PAI at the point of interest}
\usage{
.gedi_distweight(g, dist, pai, h)
}
\arguments{
\item{g}{dataframe of gedi shots returned from gedi_process}

\item{dist}{vector of distances between each gedi point and the point of interest}

\item{pai}{total PAI}

\item{h}{vegetation height (meters)}
}
\description{
Calculate the shape of the average vertical profile for a given PAI range and apply the shape to the PAI at the point of interest
}
