% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_ncepone}
\alias{.get_ncepone}
\title{Download on climate variable using the RNCEP package}
\usage{
.get_ncepone(r, tme, climvar)
}
\description{
Download on climate variable using the RNCEP package
}
