% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.hourlyall}
\alias{.hourlyall}
\title{Converts all NCEP data to hourly for a single location}
\usage{
.hourlyall(tmn, tmx, pk6, sp6, sw6, lw6, u6, v6, pr6, tmeb, tme, lat, lon)
}
\description{
Converts all NCEP data to hourly for a single location
}
