% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.inverse_distweight}
\alias{.inverse_distweight}
\title{Calculate weighted average using inverse distance weighting}
\usage{
.inverse_distweight(x, d)
}
\arguments{
\item{x}{numeric vector}

\item{d}{distances}
}
\description{
Calculate weighted average using inverse distance weighting
}
