% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.mask_water}
\alias{.mask_water}
\title{Mask out water values in rasters
lctype: land cover layer used. can be 'Copernicus', 'ESA', or 'CORINE'}
\usage{
.mask_water(r, lc, lctype)
}
\description{
Mask out water values in rasters
lctype: land cover layer used. can be 'Copernicus', 'ESA', or 'CORINE'
}
