% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.Planck}
\alias{.Planck}
\title{Used by albedo calculations - calculates wavelength specific spectral density}
\usage{
.Planck(wavelength, temperature = 5504.85)
}
\description{
Used by albedo calculations - calculates wavelength specific spectral density
}
