% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.sharpen}
\alias{.sharpen}
\title{sharpens a coarse resolution raster using a fine-resolution categorical raster}
\usage{
.sharpen(coarse, fine, msk = TRUE)
}
\description{
sharpens a coarse resolution raster using a fine-resolution categorical raster
}
