% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{era5_download}
\alias{era5_download}
\title{Downloads ERA5 climate data}
\usage{
era5_download(r, tme, credentials, file_prefix, pathout, clean = T)
}
\arguments{
\item{r}{a SpatRaster giving a template area for which to download data}

\item{tme}{a POSIXlt object giving the time period for which to download data}

\item{credentials}{a data,frame of user credentials (see vignette)}

\item{file_prefix}{character prefix to attach to the downloaded .nc files when downloaded.
The year, month, and file extension is automatically added.}

\item{pathout}{directory to which data are saved.}

\item{clean}{if TRUE, removes zip files and only saves nc files}
}
\value{
a list of the request details sent to climateData Store for all microclimate
relevant climate variables. Used by `era5_process` to process data. Files a downlaoded
to the `pathout` directory.
}
\description{
This function downloads hourly historic climate data from the Copernicus
climate data store.
}
\seealso{
[era5_process()]
}
