% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esatable}
\alias{esatable}
\title{A data frame of vegetation parameters for ESA land cover types}
\format{
a data frame with the following columns:
\describe{
 \item{Code}{Numeric code used in ESA landcover data}
 \item{Descriptor}{Description of vegetation type}
 \item{x}{Ratio of vertical to horizontal projections of leaf foliage}
 \item{gsmax}{Maximum stomatal conductances (mol / m^2 / s)}
 \item{leafd}{Leaf diameter (m)}
}
}
\usage{
esatable
}
\description{
A data frame of vegetation parameters for ESA land cover types
}
\keyword{datasets}
