% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{gedi_spline}
\alias{gedi_spline}
\title{spline gedi paiz raster to get paia at reqhgt}
\usage{
gedi_spline(gedi, h)
}
\arguments{
\item{gedi}{multilayer spatraster of gedi paia at 5 m height intervals}

\item{required}{heights of paia}
}
\description{
Returns a spatraster where each layer represents a height in the canopy with pai above the requested height
}
