% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{grid_gedi}
\alias{grid_gedi}
\title{grid gedi points}
\usage{
grid_gedi(g, template)
}
\arguments{
\item{g}{dataframe or data.table of gedi points as returned from `gedi_process`}

\item{template}{spatRaster to use as a template for gridding}
}
\description{
Grids gedi points to extent and resolution of provided template raster. 
The output raster may have NA values, particularly if a finer resolution is requested.
If NA values are present, use `lai_fill()` to fill in NA values based on landcover data
}
