% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{haduk_coastfill}
\alias{haduk_coastfill}
\title{Fills missing coastal grid cells in UK Met Office data}
\usage{
haduk_coastfill(climdata, sst, wgt = 0.5)
}
\arguments{
\item{climdata}{a list of climate data as returned by `haduk_blend`}

\item{sst}{a multi-layer SpatRast of monthly sst temperature data corresponding
to `climdata`}

\item{wgt}{an optional weighting factor between 0 and 1 indicating the sensitivity
of coastal temperatures to sea-surface temperatures.}
}
\value{
a list comprising `climarrayr` - wrapped SpatRasters of 1km hourly
climate data and `tme` - a POSIXlt object of dates and times corresponding to
data in `climarrayr`. The dataset `climarrayr` is itself a list of wrapped
SpatRasters of the following variables:
\describe{
 \item{temp}{Temperature (deg C)}
 \item{relhum}{Relative humidity (percentage)}
 \item{pres}{Sea-level pressure (kPa)}
 \item{swdown}{Downward shortwave radiation on the horizontal (W/m^2)}
 \item{difrad}{Downward diffuse radiation (W/m^2)}
 \item{lwdown}{Downward longwave radiation W/m^2}
 \item{windspeed}{wind speed at 2m (m/s)}
 \item{winddir}{wind direction (decimal degrees)}
 \item{precip}{precipitation mm}
}
}
\description{
This function infills missing coastal grid cells of the 1 km HadUK-Grid
}
\seealso{
[haduk_blend()]
}
