% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{haduk_download}
\alias{haduk_download}
\title{Downloads UK Met Office climate data}
\usage{
haduk_download(tme, credentials, pathout)
}
\arguments{
\item{tme}{a POSIXlt object giving the time period for which to download data}

\item{credentials}{a data,frame of user credentials (see vignette)}

\item{pathout}{directory to which data are saved.}
}
\value{
a data,frame detailing download details and success of all requested files
save in `pathout`.
}
\description{
This function downloads daily 1km HadUK-Grid historic temperature
and precipitation data for the entire UK.
}
\seealso{
[haduk_process()], [haduk_blend()]
}
