% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{lai_download}
\alias{lai_download}
\title{Download Leaf Area Index (LAI) data}
\usage{
lai_download(
  r,
  tme,
  reso = 10,
  pathout,
  credentials,
  pathtopython = "C:/Python/",
  download = TRUE
)
}
\arguments{
\item{r}{a SpatRaster with a coordinate reference system defined. Used
to the location for which to download data.}

\item{tme}{POSIXlt object covering time period for which data are required
(see details).}

\item{reso}{the spatial resolution of data to download. Must be either
10 or 500 (see details).}

\item{pathout}{directory to which data will be downloaded.}

\item{credentials}{a data.frame of user credentials (see vignette for
setting this up)}

\item{pathtopython}{directory in which Python is saved (see
vignette instructions for installing and setting up Python)}
}
\description{
Downloads either 10m grid resolution LAI data from WEkEO or
500m grid resolution LAI data from LPDAAC_ECS.
}
\details{
if `reso = 10` the R package `reticulate` is used to interface
with Python and download 10m resolution LAI data from the Copernicus WEkEO
site. Data are only available from 2019 to present and only for parts of
Europe (see https://www.wekeo.eu/data?view=dataset&dataset=EO%3AEEA%3ADAT%3ACLMS_HRVPP_VI
for details of coverage).If `reso = 500` 500 m grid resolution LAI data derived
from MODIS imagery are downloaded from LPDAAC_ECS using the `luna` package.
Data are available globally from 18th February 2018 onward. Multiple
files are downloaded and stored in the directory given by `pathout`. Use
[lai_mosaic()] to tile these together and [lai_fill()] to fill data gaps
caused by cloud cover. See vignette for examples.
}
\seealso{
[lai_mosaic()], [lai_fill()], [lai_downscale()],
[lai_seasonadj()]
}
