% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{lai_downscale}
\alias{lai_downscale}
\title{Downscales LAI data}
\usage{
lai_downscale(lai, lcover, veghgt = NA, aszero = NA)
}
\arguments{
\item{lai}{a coarse-resolution SpatRaster of Leaf Area Index values}

\item{lcover}{a high-resolution SpatRaster of land cover matching the
coordinate reference system of `lai`}

\item{veghgt}{optionally, a high-resolution SpatRaster of vegetation
heights matching the resolution, extent  and coordinate reference system of `lcover`}

\item{aszero}{optionally a vector of values indicating, for which land cover types,
`lai` values should automatically be set at zero}
}
\value{
a downscaled SpatRaster of Leaf Area Index values.
}
\description{
This function uses high-resolution land cover and/or vegetation
height data to downscale coarser resolution LAI.
}
\details{
if `veghgt` is not provided only land cover data are used to perform
the downscaling.
}
