% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{lai_fill}
\alias{lai_fill}
\title{Cloud cover fill lai data using landcover data}
\usage{
lai_fill(lai, lcover)
}
\arguments{
\item{lai}{a SpatRaster of Leaf Area Index values with some NAs}

\item{lcover}{a SpatRaster of land cover matching the resolution, extent and
coordinate reference system of `lai`}
}
\value{
a cloud filled SpatRaster of Leaf Area Index values.
}
\description{
This function cover fill Leaf Area Index (LAI) using land cover data
}
