% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{lai_fromndvi}
\alias{lai_fromndvi}
\title{Derive leaf area index from NDVI}
\usage{
lai_fromndvi(RGB, CIR, modisLAI, maxlai = NA)
}
\arguments{
\item{RGB}{a red-green-blue image supplied as a SpatRaster (see details).}

\item{CIR}{a colour-infrared image supplied as a SpatRaster (see details).}

\item{modisLAI}{a coarser resolution SpatRaster of MODIS-derived lai data
(see [lai_download()]).}

\item{maxlai}{optional maximum leaf area index value expected within study (see details)}
}
\value{
a SpatRaster of Leaf Area Index values
}
\description{
This function derives leaf area index data from an estimate of the
normalized difference vegetation index (NDVI).
}
\details{
`RGB` and `CIR` are normally 3-band multi-layer SpatRasters derived from
aerial photographs. It is assumed that the first band in the `RGB` image corresponds
to red and the first band in the `CIR` corresponds to near-infrared (NIR). NDVI
is then calculated as NDVI = (NIR - RED) / (NIR + RED). Since only the first
bands of `RGB` and `CIR` are used for this calaculation, `RGB` and `CIR` can
instead be single layer SpatRasters of red and near-infrared reflectance respectively.
Since aerial photographs are normally brightness and contrast adjusted during processing,
the derived NDVI values are not true NDVI. Thus the `modisLAI` SpatRaster is used to
rescale derived lai. `maxlai` is also used for rescaling, but if not supplied, is
estimated from the maximum value in `modisLAI` adjusted upwards depending on the resolution
of `RGB` and `CIR` relative to `modisLAI`. `RGB` and `CRS` must have matching
coordinate reference systems, extents and resolutions, whereas `modisLAI` is typically
of coarser resolution and reprojected if necessary.
}
