% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{lai_mosaic}
\alias{lai_mosaic}
\title{Moasic Leaf Area Index (LAI) data returned}
\usage{
lai_mosaic(r, pathin, reso = 10, msk = TRUE)
}
\arguments{
\item{r}{a SpatRaster with a coordinate reference system defined. Used
for masking and reprojecting the downloaded data.}

\item{pathin}{filepath of downloaded data. Same as `pathout` in function `lai_download`.}

\item{reso}{the spatial resolution of data download. Must be either
10 or 500 (see details under `lai_download`).}

\item{msk}{option logical indicating whether data should be masked using `r`.}
}
\description{
This function moasics Leaf Area Index (LAI) data downloaded using
`lai_download`.
}
\seealso{
[lai_download()], [lai_fill()]
}
