% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{lai_seasonadj}
\alias{lai_seasonadj}
\title{Cloud fills a multi-layer SpatRaster of LAI values by applying season effects}
\usage{
lai_seasonadj(lai)
}
\arguments{
\item{lai}{a multi-layer SpatRaster of Leaf Area Index values
with some missing values}
}
\value{
a multi-layer SpatRaster of Leaf Area Index values
without missing values
}
\description{
This function takes as an input a multi-layer SpatRaster of e.g. monthly
LAI values for a year. Missing values are interpolated temporally.
}
