% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laifeb}
\alias{laifeb}
\title{A dataset of leaf area index values}
\format{
A PackedSpatRaster object with 500 rows and 500 columns
}
\usage{
laifeb
}
\description{
A wrapped SpatRaster of leaf area index values for the Lizard Peninsula, Cornwall,
UK, for February 2023 for the area bounded by 160000, 165000, 25000, 30000
(OSGB36 / British National Grid, EPSG:27700)
}
\keyword{datasets}
