% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{micropoint_sum}
\alias{micropoint_sum}
\title{Summarize micropoint output}
\usage{
micropoint_sum(mout, sumby = "month", fun = "max")
}
\arguments{
\item{mout}{output from read_micropoint_full}

\item{sumby}{character; temporal unit over which to sum the data (can be month,
month-year, julian, julian-year, ymd, or hour)}

\item{fun}{character; function to summarize data (e.g., mean, max, min)}
}
\value{
list of dataframes; each dataframe in the list represents one of the
microclimate variables. Value = the value for the specified variable (var). The 
third column name indicates sumby function. Heights are canopy nodes
}
\description{
Summarize micropoint output
}
