% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthlylai}
\alias{monthlylai}
\title{NCEP-NOAA land-sea mask}
\format{
A PackedSpatRaster object with 94 rows and 192 columns
"landsea"
NCEP-NOAA land-sea fraction

A wrapped SpatRaster indicating what fraction of NCEP-NOAA grid cells
comprise land for the area bounded by -180.9375, 179.0625, -89.49406
(xmin, xmax, ymin, ymax) lon/lat WGS 84

A PackedSpatRaster object with 94 rows and 192 columns
"landseafrac"
A dataset of monthly leaf area index values

A wrapped multi-layer SpatRaster of monthly leaf area index values for the Lizard Peninsula, Cornwall,
UK, for Jan-Dev 2023 for the area bounded by 160000, 162000, 28000, 30000
(OSGB36 / British National Grid, EPSG:27700)

A PackedSpatRaster object with 200 rows, 200 columns
}
\usage{
monthlylai
}
\description{
A wrapped SpatRaster indicating whether NCEP-NOAA grid cells are classed as
land (1) or sea (0) for the area bounded by -180.9375, 179.0625, -89.49406
(xmin, xmax, ymin, ymax) lon/lat WGS 84
}
\keyword{datasets}
