% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{ncep_process}
\alias{ncep_process}
\title{Processes NCEP-NCAR Reanalysis climate data}
\usage{
ncep_process(
  ncepdata = NA,
  r,
  tme,
  out = "grid",
  lat = NA,
  long = NA,
  resampleout = FALSE
)
}
\arguments{
\item{ncepdata}{list of climate variables as returned by `ncep_download` - data
downloaded if not supplied}

\item{r}{a SpatRaster object as a template or downloading data or for resampling
and masking outputs if specified.}

\item{tme}{POSIXlt object for which processed data are required. Must overlap
with that supplied in `ncepdata`.}

\item{out}{one of `grid` or `point` specifying whether to return a list of
arrays of climate data or a data.frame for a single point location.}

\item{lat}{latitude (decimal degrees) - alternative to supplying a template
raster if data are required for a point location.}

\item{resampleout}{optional logical indicating whether to resample and mask outputs
to match `r`.}

\item{lon}{longitude (decimal degrees) - alternative to supplying a template
raster if data are required for a point location.}
}
\value{
if `out = "point"` then a a data.frame of hourly weather comprising
the following columns:
\describe{
 \item{obs_time}{POSIXlt object of dates and times}
 \item{temp}{Temperature (deg C)}
 \item{relhum}{Relative humidity (percentage)}
 \item{pres}{Sea-level pressure (kPa)}
 \item{swdown}{Downward shortwave radiation on the horizontal (W/m^2)}
 \item{difrad}{Downward diffuse radiation (W/m^2)}
 \item{lwdown}{Downward longwave radiation W/m^2}
 \item{windspeed}{wind speed at 2m (m/s)}
 \item{winddir}{wind direction (decimal degrees)}
 \item{precip}{precipitation mm}
}

If `out = "grid"`a list of wrapped multilayer SpatRasters of each
of the above climate variables (i.e. `tme` not returned - same as input).
}
\description{
This function processes six-hourly climate data from the NCEp-NCAR
reanalysis project for use with `microgrid` or `micropoint`
}
\seealso{
[ncep_download()]
}
