% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{padcalc}
\alias{padcalc}
\title{Calculate plant area density profiles from airborne lidar data}
\usage{
padcalc(las, hres, dz, k)
}
\arguments{
\item{las}{lidR object. Should be normalized and classified with ground = 2}

\item{hres}{horizontal resolution}

\item{dz}{vertical resolution}

\item{k}{extinction coefficient}
}
\value{
spatRaster of PAD profiles. Names of the layers indicate the vertical 
bin from ground to canopy
}
\description{
This function takes a lidR point cloud and calculates PAD profiles at the 
specified horizontal and vertical resolutions. Results differ slightly from
PAD profiles calculated in leafR and canopylazR because ground points are 
not considered to contribute to PAD in the bottom-most layer. Doing so
the ground does not contribute to vegetation in microclimate models. This
function also uses terra rather than raster, which the older leafR and 
canopylazR depend on.
}
\examples{
las = readLas(lasnorm.laz)
pad <- padcalc(las, hres = 5, dz = 1, k = 0.5)

}
