% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{read_micropoint_full}
\alias{read_micropoint_full}
\title{read micropoint full}
\usage{
read_micropoint_full(
  fname,
  microvars = c("tair", "tleaf", "relhum"),
  read_inputs = NULL
)
}
\arguments{
\item{fname}{.h5 file name to read}

\item{microvars}{which microclimate variables should be read from the output.
Can be any names including in output of RunModelFull and that were saved when 
running `write_micropoint_full()`.}

\item{read_inputs}{logical; should inputs be read in addition to microclimate outputs}
}
\value{
description returns list of microclimate variables, heights of nodes, 
lat, lon, time, and input parameters if specified. Microclimate arrays are organized
as described in `write_micropoint_full()`
}
\description{
Read output from `write_micropoint_full()`
}
