% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{reflectance_calc}
\alias{reflectance_calc}
\title{Derives leaf and ground reflectance}
\usage{
reflectance_calc(
  alb,
  lai,
  x,
  plotprogress = TRUE,
  maxiter = 50,
  tol = 0.001,
  bwgt = 0.5
)
}
\arguments{
\item{alb}{a SpatRaster of albedo values.}

\item{lai}{a SpatRaster of leaf afrea index values.}

\item{x}{a SpatRaster of leaf inclination coefficient values as returned by [x_calc()].}

\item{plotprogress}{optional logical indicating whether or not to keep track ofprogress by plotting interim results.}

\item{maxiter}{maximum number of iterations}

\item{tol}{convergence tolerance}

\item{bwgt}{value between zero and one indicating by how much to back-weight results during iteration. HIgher values
ensure more stable convergence, but may require more iteration to achieve convergence.}
}
\value{
a list comprising SpatRasters of leaf and ground reflectance
}
\description{
Derives leaf and ground reflectance from albedo.
}
