% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{soildata_download}
\alias{soildata_download}
\title{Downloads soil data}
\usage{
soildata_download(r, pathdir = getwd(), deletefiles = TRUE)
}
\arguments{
\item{r}{a SpatRaster object used as a template for downloading data (see details).}

\item{deletefiles}{optional logical indicating whether to delete the original downloaded files for each depth}

\item{pathdi}{directory to which data are temporarily or permenently downloaded (default working directory).}
}
\value{
a multi-layer SpatRaster of depth-weighted soil physical-properties:
#' \describe{
 \item{bdod}{Bulk density in cg/cm^3}
 \item{clay}{Mass fraction of clay particles g/kg}
 \item{sand}{Mass fraction of sand particles g/kg}
 \item{silt}{Mass fraction of silt particles g/kg}
 }
}
\description{
Downloads data on soil properties from the https://soilgrids.org/
}
\seealso{
[soildata_downscale()]
}
