% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{soildata_downscale}
\alias{soildata_downscale}
\title{Downscales soil data using landcover data}
\usage{
soildata_downscale(soildata, landcover, water = 80)
}
\arguments{
\item{soildata}{a multi-layer SpatRaster of soil physical properties as returned by `soildata_download`.}

\item{landcover}{a high-resolution SpatRaster of landcover as returned by e.g. `lcover_download`}

\item{water}{value in land cover corresponding to water. Used to clean soil data}
}
\value{
a multi-layer SpatRaster of downscaled depth-weighted soil physical-properties
}
\description{
Fills NAs and Downscales soil data derived using `soildata_download` using
landcover data.
}
\seealso{
[soildata_download()]
}
