% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{soildata_gettype}
\alias{soildata_gettype}
\title{Derives a soil type from soil data}
\usage{
soildata_gettype(soildata)
}
\arguments{
\item{soildata}{a multi-layer SpatRaster of soil physical properties as returned by `soildata_download`.
or `soildata_downscale`.}
}
\value{
a SpatRaster of soil types one of:
\describe{
 \item{1}{Sand}
 \item{2}{Loamy sand}
 \item{3}{Sandy loam}
 \item{4}{Loam}
 \item{5}{Silt loam}
 \item{6}{Sandy clay loam}
 \item{7}{Clay loam}
 \item{8}{Silty clay loam}
 \item{9}{Sandy clay}
 \item{10}{Silty clay}
 \item{11}{Clay}
}
}
\description{
Derives a numerical code corresponding to likely soil type from bulk density and the
sand, clay and silt mass fraction.
}
