% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{sst_download}
\alias{sst_download}
\title{Downloads daily sea-surface temperature data}
\usage{
sst_download(r, tme, resampleout = "FALSE", nafill = "FALSE")
}
\arguments{
\item{r}{a SpatRaster giving a template area for which to download data}

\item{tme}{a POSIXlt object giving the time period for which to download data}
}
\value{
a multi-layer SpatRaster of daily sea-surface temperature data
}
\description{
This function downloads daily 0.25 degree grid resolution sea-surface
temperature data from the NOAA National Centers for Environmental Information
}
\seealso{
[haduk_blend()]
}
