% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{topo_process}
\alias{topo_process}
\title{Processes topographic data}
\usage{
topo_process(dem, lc, lctype)
}
\arguments{
\item{dem}{The elevation raster from dem_download}

\item{lc}{Landcover layer}

\item{lctype}{type of landcover layer. Can be 'CORINE', 'ESA', or 'Copernicus' (see descriptions in lc_download)}
}
\value{
a multi-layer SpatRaster of elevation, slope (degrees), and aspect (degrees)
}
\description{
Processes elevation data to produce a topography raster containing elevation, slope, and aspect. Masks out water
}
\seealso{
[dem_download()]
}
