% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{ukcp18_download}
\alias{ukcp18_download}
\title{Download UKCP18 climate data, including future climate}
\usage{
ukcp18_download(
  tme,
  cedatoken,
  pathout,
  collection = c("land-gcm", "land-rcm"),
  domain = c("uk", "eur", "global"),
  rcp = c("rcp85", "rcp26"),
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  silent = FALSE
)
}
\arguments{
\item{tme}{POSIXlt object indicating for which time period data are downloaded}

\item{cedatoken}{ceda access token as provided by \href{https://services-beta.ceda.ac.uk/account/token/?_ga=2.221117848.1874942177.1727205775-405001952.1710779653}{Ceda access token generator}}

\item{pathout}{directory to which to save data}

\item{collection}{text string defining UKCP18 collection, either `land-gcm`
(60 km grid resolution) or `land-rcm` (12 km grid resolution)}

\item{domain}{text string defining UKCP18 domain, either 'uk' (United Kingdom)
or 'eur' (Europe) or for `land-gcm`, `global` (global extent)}

\item{rcp}{text string of RCP scenario to be downloaded (either `rcp85` or `rcp26`)}

\item{member}{vector of strings defining the climate model member to be downloaded. Available members vary between UKCP18 collections.}

\item{silent}{optional logical indicating whether to print which files are being downloaded.}
}
\description{
The function downloads specified data from ftp.ceda.ac.uk to a specified directory
}
