% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{ukcp18_process}
\alias{ukcp18_process}
\title{Process UKCP18 data converting to daily data}
\usage{
ukcp18_process(
  r,
  tme,
  pathin,
  collection = c("land-gcm", "land-rcm"),
  domain = c("uk", "eur", "global"),
  rcp = c("rcp85", "rcp26"),
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  resampleout = FALSE
)
}
\arguments{
\item{r}{SpatRaster giving extent of returned data. If resampleout is set to TRUE,
returned data are projected / resampled to match r,}

\item{tme}{POSIXlt object indicating for which time period data are downloaded}

\item{pathin}{directory to which data are saved.}

\item{collection}{text string defining UKCP18 collection, either `land-gcm`
(60 km grid resolution) or `land-rcm` (12 km grid resolution)}

\item{domain}{text string defining UKCP18 domain, either 'uk' (United Kingdom)
or 'eur' (Europe) or for `land-gcm`, `global` (global extent)}

\item{rcp}{text string of RCP scenario to be downloaded (either `rcp85` or `rcp26`)}

\item{member}{strings defining the climate model member downloaded.}

\item{resampleout}{optional logical indicating whether to reproject / resample output to match `r`}
}
\description{
The function reads in downloaded UKCP18 data, crops out the right area
and time slice and conbverts data form 360 days per year to true daily data.
}
