% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{write_micropoint_full}
\alias{write_micropoint_full}
\title{Write timeseries profiles from RunModelFull}
\usage{
write_micropoint_full(
  mout,
  fname,
  h,
  st_time,
  en_time,
  microout = c("tair", "tleaf", "relhum"),
  vegp = NULL,
  soilp = NULL,
  macroclim = NULL,
  zref = NULL,
  Lfrac = NULL,
  paii,
  overwrite = T
)
}
\arguments{
\item{mout}{output of RunModelFull}

\item{fname}{file name (must be .h5)}

\item{h}{heights of canopy nodes in mout (represent top of each bin)}

\item{st_time}{start time}

\item{en_time}{end time - assumes time step between start and end times is 1 hour}

\item{microout}{vector of parameters to save; can be any of the names of mout (e.g., tair, tleaf, relhum)}

\item{vegp}{if not NULL, writes vegp inputs}

\item{soilp}{if not NULL, writes soilc inputs}

\item{macroclim}{if not null writes macroclimate data inputs}

\item{zref}{if not NULL writes zref parameter}

\item{Lfrac}{if not NULL writes Lfrac inputs}

\item{paii}{if not NULL writes paii inputs}
}
\value{
writes an h5 file with three groups. 1) microclim/ contains modelled
 microclimate arrays. Writes the variables specified in microout. Columns represent
 time (hour), and rows represent height in the canopy. Columns are ordered from
 first(left) to last (right), and heights are ordered from top of canopy (1)
 to bottom of canopy (nrow (microclim)), Therefore microclim[[1]][1,1]
 represents the top of the canopy at the first time step and microclim[[1]][nrow(out),1]
 represents the bottom of the canopy at the first time step.
 The microclim group contains four attributes 1) height of each canopy node (top to bottom),
 start time (st_time), end time (en_time), and time step in hours. Time variables
 allow recreation of the obs_time vector used in the models.
 Group 2) inputs - if inputs are provided, they are written the the file
 Group 3) metadata - contains, lat, lon, time_zone
}
\description{
Write output from RunModelFull
}
