% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{x_calc}
\alias{x_calc}
\title{Calculate leaf inclination coefficient}
\usage{
x_calc(landcover, lctype = "ESA")
}
\arguments{
\item{landcover}{a SpatRaster of landcover class as returned by [lcover_download()].}

\item{lctype}{one of `ESA`, `CORINE`, or `Copernicus` indicating the source of `landcover`.}
}
\value{
a SpatRaster leaf inclination coefficient values
}
\description{
Estim,ates typical values of the ratio of vertical to horizontal
projections of leaf foliage from land cover data,
}
