% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{climpoint_extract}
\alias{climpoint_extract}
\title{extracts climate data for a point location}
\usage{
climpoint_extract(
  climdata,
  x,
  y,
  aslatlong = FALSE,
  altcorrect = 2,
  dtmc,
  elev,
  proj = FALSE,
  distweight = T
)
}
\arguments{
\item{climdata}{a list of SpatRasters of climate data as returned by haduk_coastfill or
any other of the climate data process functions.}

\item{x}{easting or longitude}

\item{y}{northing or latitude}

\item{aslatlong}{optional logical indicating whether `x` and `y` are longitude
and latitude (TRUE) or the easting and northing in the same coordinate reference
system as `climdata`}

\item{altcorrect}{0=no altitudinal correction; 1=temperature dependent lapse rate correction;
humidity dependent lapse rate correction}

\item{dtmc}{raster of era5 elevation data (optional; only needed if altcorrect>0)}

\item{elev}{raster of elevation or numeric elevation (optional; only needed if altcorrect>0)}

\item{proj}{set TRUE if x,y, and climate are in projected coordinate system}

\item{distweight}{if TRUE applies bilinear interpolation using 4 nearest neighbors to point of interest}
}
\description{
This function extracts climate data from a climate arrays for any
given point location returning a data,frame as used by the `micropoint` package
}
