% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{climpoint_extract}
\alias{climpoint_extract}
\title{extracts climate data for a point location}
\usage{
climpoint_extract(climdata, x, y, aslatlong = FALSE)
}
\arguments{
\item{climdata}{a list of SpatRasters of climate data as returned by haduk_coastfill or
any other of the climate data process functions.}

\item{x}{easting or longitude}

\item{y}{northing or latitude}

\item{aslatlong}{optional logical indicating whether `x` and `y` are longitude
and latitude (TRUE) or the easting and northing in the same coordinate reference
system as `climdata`}
}
\description{
This function extracts climate data from a climate arrays for any
given point location returning a data,frame as used by the `micropoint` package
}
