% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{create_veggrid}
\alias{create_veggrid}
\title{Create vegetation inputs for grid model}
\usage{
create_veggrid(landcover, vhgt, lai, refldata, lctype = "ESA")
}
\arguments{
\item{landcover}{a SpatRaster of landcover class as returned by [lcover_download()].}

\item{vhgt}{a SpatRaster of vegetation height as returned by [vegheight_download()].}

\item{lai}{a SpatRaster of Leaf Area Index values.}

\item{refldata}{a list of ground and leaf reflectance values as returned by [reflectance_calc()].}

\item{lctype}{one of `ESA` or `CORINE` indicating the source of `landcover`. Alternatively a data.frame
matching vegetation parameters to landcover types following the format of the inbuilt datasets `esatable`
and `corinetable`.}
}
\value{
an object of class `vegparams`, namely a list of the following objects:
\describe{
  \item{pai}{a PackedSpatRaster of plant area index values}
  \item{hgt}{a PackedSpatRaster object of vegetation heights (m)}
  \item{x}{a PackedSpatRaster object of the ratio of vertical to horizontal projections of leaf foliage}
  \item{gsmax}{a PackedSpatRaster object maximum stomatal conductances (mol / m^2 / s)}
  \item{leafr}{a PackedSpatRaster object of leaf reflectance values}
  \item{clump}{a PackedSpatRaster of monthly values between 0 and 1 indicating the fraction of radiation passing through larger gaps in the canopy}
  \item{leafd}{a PackedSpatRaster object of leaf diameters (m)}
  \item{leaft}{a PackedSpatRaster object of leaf transmittance}
}
}
\description{
Creates an object of class vegparams used as an input to `microclimf`
or `microgrid`.
}
