% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{create_vegpoint}
\alias{create_vegpoint}
\title{Create vegetation inputs for point model}
\usage{
create_vegpoint(
  landcover,
  vhgt,
  lai,
  refldata,
  lctype = "ESA",
  lat = NA,
  long = NA
)
}
\arguments{
\item{landcover}{a SpatRaster of landcover class as returned by [lcover_download()].}

\item{vhgt}{a SpatRaster of vegetation height as returned by [vegheight_download()].}

\item{lai}{a SpatRaster of Leaf Area Index values.}

\item{refldata}{a list of ground and leaf reflectance values as returned by [reflectance_calc()].}

\item{lctype}{one of `ESA` or `CORINE` indicating the source of `landcover`. Alternatively a data.frame
matching vegetation parameters to landcover types following the format of the inbuilt datasets `esatable`
and `corinetable`.}

\item{lat}{optionally, latitude (decimal degrees) - see details.}

\item{long}{optionally, longitude (decimal degrees) - see details.}
}
\value{
an object of class `vegparams`, namely a list of the following objects:
\describe{
  \item{pai}{Plant area index value}
  \item{hgt}{Vegetation height (m)}
  \item{x}{Ratio of vertical to horizontal projection of leaf foliage}
  \item{gsmax}{Maximum stomatal conductances (mol / m^2 / s)}
  \item{leafr}{Leaf reflectance}
  \item{clump}{Value between 0 and 1 indicating the fraction of radiation passing through larger gaps in the canopy}
  \item{leafd}{Leaf diameter (m)}
  \item{leaft}{Leaf transmittance}
}
}
\description{
Creates an object of class vegparams used as an input to `micropoint`.
}
\details{
If `lat` and `long` are unspecified, point data are selelected for
the location and the centre of supplied `landcover`.
}
