% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.getL2Bprofile}
\alias{.getL2Bprofile}
\title{Get L2B profile including beam number, shot number, quality flag, time, lat, lon, elevation, rh100 (canopy height), pai (total pai), and pai_z. Adapted from rGEDI}
\usage{
.getL2Bprofile(level2b, clean = T)
}
\description{
Get L2B profile including beam number, shot number, quality flag, time, lat, lon, elevation, rh100 (canopy height), pai (total pai), and pai_z. Adapted from rGEDI
}
