% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{gedi_download}
\alias{gedi_download}
\title{download GEDI data for microclimate models}
\usage{
gedi_download(aoi, outdir, creds, overwrite = F, clean = T)
}
\arguments{
\item{aoi}{spatRaster or spatVector defining the area of interest. Downloads 
orbits that intersect with the extent.}

\item{outdir}{directory to save files}

\item{creds}{credentials dataframe as described in microclimdta vignette}

\item{overwrite}{should files be overwritten}

\item{clean}{if T, h5 files are deleted}
}
\description{
downloads and saves quality checked vertical profiles of PAI as csv
}
