% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{gedi_process}
\alias{gedi_process}
\title{Process GEDI data for use in point-based microclimate models}
\usage{
gedi_process(
  l2b,
  aoi,
  powerbeam = TRUE,
  yr = NULL,
  mth = NULL,
  night = TRUE,
  leafon = FALSE
)
}
\arguments{
\item{l2b}{character vector of hd5 files of GEDI L2B data (downloaded using gedi_download)}

\item{aoi}{spatRaster or spatVector to crop gedi data}

\item{powerbeam}{default TRUE; if TRUE filters to only power beams; if FALSE includes power and coverage beams}

\item{yr}{optional; 4 digit year (YYYY) to filter GEDI data}

\item{mth}{optional; 2 digit month (MM) to filter GEDI data}

\item{night}{filter out daytime shots (default = TRUE)}

\item{leafon}{only keep shots during leaf on period (default=FALSE)}
}
\value{
a dataframe of vegetation information with the following values
\describe{
  \item{pai}{Total plant area index value}
  \item{pai_z_}{plant area index between z1 to z2 m above ground}
}
}
\description{
Extracts data from the L2B profile and calculates vertical PAI profile. The sum of the vertical PAI profile = total PAI
}
