% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\name{lcover_download}
\alias{lcover_download}
\title{Download Landcover data}
\usage{
lcover_download(
  r,
  type = "ESA",
  year = 2018,
  GoogleDrivefolder,
  pathtopython,
  projectname = NA,
  silent = FALSE
)
}
\arguments{
\item{r}{a SpatRaster with a coordinate reference system defined. Used
to the location for which to download data.}

\item{type}{one of `ESA` or `CORINE`}

\item{year}{year for which land cover data are required. Ignored if `type` set
to `ESA` as 2023 data will be downloaded. Options are 1986 to 2018 for `CORINE`.}

\item{GoogleDrivefolder}{name of folder on your Google drive to which data will
be downloaded.}

\item{pathtopython}{directory in which Python is saved (see
vignette instructions for installing and setting up Python)}

\item{projectname}{the name of your Google Earth Engine project (see
vignette instructions.}

\item{silent}{optional logical indicating whether or not to inform user of download
progress.}
}
\description{
Downloads either 10m grid resolution ESA or 100m grid resolution
CORINE landcover data from Google Earth Engine. The latter contains more habitat
classes.
}
\details{
See vignette for example of how to use ths function.
}
